requirements.txt
# Core dependencies
PyYAML>=6.0.1
tqdm>=4.66.1
mutagen>=1.47.0
ffmpeg-python>=0.2.0
colorama>=0.4.6

# Audio processing
ffmpeg>=4.4.0  # System dependency
ffprobe>=4.4.0  # System dependency

# Database
sqlite3  # Built-in Python module

# Optional dependencies
requests>=2.31.0  # For SongLink module
beautifulsoup4>=4.12.0  # For web scraping 